+++ wn/_core.py
@@ -12,6 +12,7 @@ from typing import (
     Iterator,
 )
 import warnings
+import textwrap
 
 import wn
 from wn._types import (
@@ -204,6 +205,42 @@ class Lexicon(_DatabaseEntity):
         """
         return [_to_lexicon(get_lexicon(rowid))
                 for rowid in get_lexicon_extensions(self._id, depth=depth)]
+
+    def describe(self, full: bool = True) -> str:
+        """Return a formatted string describing the lexicon.
+
+        The *full* argument (default: :python:`True`) may be set to
+        :python:`False` to omit word and sense counts.
+
+        """
+        _id = self._id
+        substrings: List[str] = [
+            f'{self.specifier()}',
+            f'  Label  : {self.label}',
+            f'  URL    : {self.url}',
+            f'  License: {self.license}',
+        ]
+        if full:
+            substrings.extend([
+                f'  Words  : {_desc_counts(find_entries, _id)}',
+                f'  Senses : {sum(1 for _ in find_senses(lexicon_rowids=(_id,)))}',
+            ])
+        substrings.extend([
+            f'  Synsets: {_desc_counts(find_synsets, _id)}',
+            f'  ILIs   : {sum(1 for ili, *_ in find_ilis(lexicon_rowids=(_id,))):>6}',
+        ])
+        return '\n'.join(substrings)
+
+
+def _desc_counts(query: Callable, lexid: int) -> str:
+    count: Dict[str, int] = {}
+    for _, pos, *_ in query(lexicon_rowids=(lexid,)):
+        if pos not in count:
+            count[pos] = 1
+        else:
+            count[pos] += 1
+    subcounts = ', '.join(f'{pos}: {count[pos]}' for pos in sorted(count))
+    return f'{sum(count.values()):>6} ({subcounts})'
 
 
 class _LexiconElement(_DatabaseEntity):
@@ -1173,6 +1210,18 @@ class Wordnet:
         """
         iterable = find_ilis(status=status, lexicon_rowids=self._lexicon_ids)
         return [ILI(*ili_data) for ili_data in iterable]
+
+    def describe(self) -> str:
+        """Return a formatted string describing the lexicons in this wordnet."""
+        substrings = ['Primary lexicons:']
+        for lex in self.lexicons():
+            substrings.append(textwrap.indent(lex.describe(), '  '))
+        expanded = self.expanded_lexicons()
+        if expanded:
+            substrings.append('Expand lexicons:')
+            for lex in self.expanded_lexicons():
+                substrings.append(textwrap.indent(lex.describe(full=False), '  '))
+        return '\n'.join(substrings)
 
 
 def _to_lexicon(data) -> Lexicon: